Switch to ImageMagick pixelCache.
This avoids numerous calls to pixelColor which will try to modify small
parts of the picture in a lazy way, which makes it all very slow when
resizing.

Please note there appears to be no official API for accessing the actual
pixels in the pixel cache, so we need to do it the dirty way.